Account for Shipping being displayed later in checkout flow
For some reason these specs started failing due to the shipping charge being shown later than the spec expected.
Unsure if this is a bug I inadvertently introduced into the spec or if Spree has changed its behavior, but more likely the former.
Test Cases 2b and 6 are passing again.